+++ migrations/20250307152458_alter_default_wallet_balance.js
@@ -1,0 +1,19 @@
+/**
+ * @param { import("knex").Knex } knex
+ * @returns { Promise<void> }
+ */
+exports.up = function (knex) {
+  return knex.schema.withSchema('users').alterTable('wallet', tbl => {
+    tbl.bigint('balance').defaultTo(0).alter();
+  });
+};
+
+/**
+ * @param { import("knex").Knex } knex
+ * @returns { Promise<void> }
+ */
+exports.down = function (knex) {
+  return knex.schema.withSchema('users').alterTable('wallet', tbl => {
+    tbl.bigint('balance').defaultTo(10000).alter();
+  });
+};