+++ src/components/ui/InputGroup.tsx
@@ -1,3 +1,3 @@
 export function InputGroup({ children }) {
-  return <div className='flex flex-col gap-2 w-full'>{children}</div>;
+  return <div className='flex flex-col gap-1 w-full'>{children}</div>;
 }