+++ src/features/users/register/schemas/registerCredentialsSchema.ts
@@ -1,13 +1,20 @@
+import { passwordSchema } from '@/schemas/passwordSchema';
 import z from 'zod';
 
 export const registerCredentialsSchema = z
   .object({
     email: z.string().email().nonempty().trim(),
-    password1: z.string().min(8).max(16).nonempty(),
-    password2: z.string().min(8).max(16).nonempty(),
+    password1: passwordSchema,
+    password2: passwordSchema,
   })
   .strict()
-  .refine(value => {
-    const { password1, password2 } = value;
-    return password1 === password2;
-  });
+  .refine(
+    value => {
+      //Make sure the passwords are the same.
+      const { password1, password2 } = value;
+      return password1 === password2;
+    },
+    {
+      message: 'password_mismatch',
+    }
+  );